refactor(categories): migrate the useGetAllCategoriesQuery to tanstack query
